Adds formula for simavr, another AVR simulator.
I have only tested this on my local machine so far, so there might still be
bugs lurking around somewhere. Compiling is not really a problem, except that
it seems to be hard to convince the Makefile to accept ENV variables (make
needs to be called with -e switch), but I think I got it working.